fix: Remove redundant logic in tts exampels
